Code/agentic-ml-builder/src/requirements.txt
# Core Dependencies
openai==1.57.1
azure-ai-inference==1.0.0b5
azure-identity==1.19.0
azure-storage-blob==12.23.1
azure-mgmt-machinelearningservices==2.0.0
python-dotenv==1.0.1

# ML/Data Science
torch==2.5.1
torchvision==0.20.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.2
tensorboard==2.18.0

# RAG and Vector Store
chromadb==0.5.23
sentence-transformers==3.3.1
langchain==0.3.13
faiss-cpu==1.9.0

# MCP
mcp==0.9.0
httpx==0.28.1

# Testing and Validation
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
docker==7.1.0
flake8==7.1.1

# API and File Handling
requests==2.32.3
aiofiles==24.1.0
pyyaml==6.0.2
jinja2==3.1.5

# OneDrive Integration
msal==1.31.1
msgraph-sdk==1.12.0

# CLI and Utilities
click==8.1.8
rich==13.9.4
tqdm==4.67.1

# Azure SDK
azure-ai-ml==1.22.2
azureml-core==1.59.0